Add SetFastTravelWaitTimeout
This allows scripts waiting on FastTravelPrompt to set a timeout period
before fast traveling. This can be used if the script is delayed and
having issues setting a new target. FastTravel will wait until either a
SetFastTravelTarget is set or the timeout is exceeded.

Delays of 0.0 will remove the timeout. User may notice the delay as it
will pause the UI while clicking confirm. Delay is in ms.